chore(test): move tests to match new file structure
