+++ test/core/di/povider.spec.ts
@@ -1,18 +1,18 @@
 import {expect} from 'chai';
-import {provide} from "../../src/di/provider";
-import {Inject,Injectable} from "../../src/di/decorators";
-import {InjectMetadata,OptionalMetadata,HostMetadata} from "../../src/di/metadata";
-import {Component,Directive} from "../../src/directives/decorators";
-import {Pipe} from '../../src/pipes/decorators'
-import {noop} from "../../src/facade/lang";
-import {_areAllDirectiveInjectionsAtTail} from "../../src/di/provider";
-import {_getTokenStringFromInjectable} from '../../src/di/provider';
-import {OpaqueToken} from '../../src/di/opaque_token';
-import {_extractToken} from '../../src/di/provider';
-import {_dependenciesFor} from '../../src/di/provider';
-import {Host} from '../../src/di/decorators';
-import {isFunction} from '../../src/facade/lang';
-import {ParamMetaInst} from '../../src/di/provider';
+import {provide} from '../../../src/core/di/provider';
+import {Inject,Injectable} from '../../../src/core/di/decorators';
+import {InjectMetadata,OptionalMetadata,HostMetadata} from '../../../src/core/di/metadata';
+import {Component,Directive} from '../../../src/core/directives/decorators';
+import {Pipe} from '../../../src/core/pipes/decorators'
+import {noop} from '../../../src/facade/lang';
+import {_areAllDirectiveInjectionsAtTail} from '../../../src/core/di/provider';
+import {_getTokenStringFromInjectable} from '../../../src/core/di/provider';
+import {OpaqueToken} from '../../../src/core/di/opaque_token';
+import {_extractToken} from '../../../src/core/di/provider';
+import {_dependenciesFor} from '../../../src/core/di/provider';
+import {Host} from '../../../src/core/di/decorators';
+import {isFunction} from '../../../src/facade/lang';
+import {ParamMetaInst} from '../../../src/core/di/provider';
 
 describe( `di/provider`, ()=> {
