+++ test/index.ts
@@ -1,0 +1,28 @@
+//import { expect } from 'chai';
+
+//import './test/host.spec';
+//import './test/inject.spec';
+//import './test/injectable.spec';
+//import './test/util.spec';
+//import './test/directives.spec';
+//import './test/bootstrap.spec';
+//import './test/pipe.spec';
+//import './test/life_cycle.spec';
+//import './test/providers.spec';
+
+import './core/di/decorators.spec';
+import './core/di/povider.spec';
+import './core/util/decorators.spec';
+import './core/reflection/reflection.spec';
+import './core/linker/pipe_resolver.spec';
+import './core/pipes/pipe_provider.spec';
+import './core/linker/directive_lifecycles_reflector.spec';
+import './core/linker/directive_resolver.spec';
+import './core/directives/directive_provider.spec';
+
+import './facade/lang.spec';
+import './facade/primitives.spec';
+import './facade/collections.spec';
+
+describe( 'ng-metadata', ()=> {
+} );
